Separate Popover Backdrop in separate component for code clarity.
